% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addEmptyColumnsBeforeSpecificIndex}
\alias{addEmptyColumnsBeforeSpecificIndex}
\title{add column before specific index}
\usage{
addEmptyColumnsBeforeSpecificIndex(index, df)
}
\arguments{
\item{index}{column index you want to add empty row before}

\item{df}{data frame}
}
\description{
add column before specific index
}
