% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addEmptyRowsBeforeSpecificIndex}
\alias{addEmptyRowsBeforeSpecificIndex}
\title{add row before specific index}
\usage{
addEmptyRowsBeforeSpecificIndex(index, df)
}
\arguments{
\item{index}{row index you want to add empty row before}

\item{df}{data frame}
}
\description{
add row before specific index
}
