% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable-format-tidier.R
\name{addTableName}
\alias{addTableName}
\title{Function To add table name}
\usage{
addTableName(tableName, align = "left", colwidths)
}
\arguments{
\item{align}{align of the table name. It includes left, right and center}

\item{colwidths}{colwidth. It saves time to use col_keys length}

\item{x}{a flextable object}
}
\value{
return a flextable that meets the need of academic paper
}
\description{
Function To add table name
}
\details{
In academic paper, we use three line table. This is a simpilified version to export the table to save time for group meeting.
}
\examples{
# examples


iris \%>\%
  flextable() \%>\%
  themeGm()
}
