% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c218Tools-package.R
\docType{package}
\name{c218Tools-package}
\alias{c218Tools}
\alias{c218Tools-package}
\title{c218Tools: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
