% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cbnRegLst}
\alias{cbnRegLst}
\title{combine the result from regDisplay}
\usage{
cbnRegLst(exposureLength, ModelNum, regList)
}
\arguments{
\item{exposureLength}{length of target}

\item{regList}{regression list}

\item{modelNum}{model number}
}
\description{
combine the result from regDisplay
}
