% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuteArt.R
\name{cuteArt}
\alias{cuteArt}
\title{Print a cute ASCII art}
\usage{
cuteArt(animal = c("cat", "bear", "rabbit", "dog", "panda"))
}
\arguments{
\item{animal}{which art to print, one of "cat", "bear", "rabbit", "dog",
"panda". Default "cat".}
}
\value{
invisibly returns the art as a character string.
}
\description{
Print a cute ASCII art to the console. Just for fun / testing.
}
\examples{
cuteArt()
cuteArt("bear")
cuteArt("panda")
}
