% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTargetLevels.R
\name{detectOutcomeLevels}
\alias{detectOutcomeLevels}
\title{detect outcome levels}
\usage{
detectOutcomeLevels(outcome, data)
}
\arguments{
\item{outcome}{your outcome}

\item{data}{your data}
}
\description{
detect outcome levels
}
