% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation-mediate.R
\name{extract_mediation_summary}
\alias{extract_mediation_summary}
\title{extract summary mediation}
\usage{
extract_mediation_summary(x)
}
\arguments{
\item{x}{mediate or summary.mediate object}
}
\value{
\item{smat}{regression table of mediation}
}
\description{
extract summary mediation
}
\details{
This function was from https://stackoverflow.com/questions/41582486/how-to-convert-r-mediation-summary-to-data-frame. Thanks for the  user hrbrmstr's answer on Jan 11, 2017 at 4:39
}
\examples{
# demo

}
