% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable-format-tidier.R
\name{formatTable}
\alias{formatTable}
\title{Function To set table style}
\usage{
formatTable(x, tableName, align = "left", colwidths = NULL)
}
\arguments{
\item{x}{a flextable object}

\item{tableName}{the table name (caption) to add on top of the table}

\item{align}{align of the table name. It includes left, right and center}

\item{colwidths}{colwidth. It saves time to use col_keys length. If NULL (default), it spans all columns.}
}
\value{
return a flextable that meets the need of academic paper
}
\description{
Function To set table style
}
\details{
In academic paper, we use three line table. This is a simpilified version to export the table to save time for group meeting.
}
\examples{
# examples

iris |>
  flextable::flextable() |>
  themeGm()
}
