% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaGenerator.R
\name{genForm}
\alias{genForm}
\title{Formula generator}
\usage{
genForm(outcome, exposure, covariate, sensitivity = F)
}
\arguments{
\item{outcome}{outcome of your regression.}

\item{exposure}{exposure you are interested in.}

\item{covariate}{covariate you want to adjust. If you want to generate formula with several batch of variables to adjust. please provide a list of vectors}

\item{sensitivity}{whether to generate forumla for sensitivity analysis. Default is FALSE.}
}
\description{
Formula generator
}
\examples{
# examples

outcome <- "outcome"
exposure <- "exposure"
covariate <- c("a", "b")
genForm(outcome = outcome, exposure = exposure, covariate = covariate)


outcome <- "outcome"
exposure <- "exposure"
covariate <- list(c("a", "b"), c("c", "d"))
genForm(outcome = outcome, exposure = exposure, covariate = covariate)

outcome <- "outcome"
exposure <- c("exposure", "exposure2")
covariate <- list(c("a", "b"), c("c", "d"))
form <- lapply(exposure, function(x){
  genForm(outcome = outcome, exposure = x, covariate = covariate)
})

}
