% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTableOne.R
\name{makeNamesTableOneExcel}
\alias{makeNamesTableOneExcel}
\title{try to clean names of tableone}
\usage{
makeNamesTableOneExcel(x)
}
\arguments{
\item{x}{An object returned from tableone package}
}
\description{
try to clean names of tableone
}
