% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTableOne.R
\name{makeNamesTableOneLatex}
\alias{makeNamesTableOneLatex}
\title{try to clean names of tableone}
\usage{
makeNamesTableOneLatex(x)
}
\arguments{
\item{x}{An object returned from tableone package}
}
\description{
try to clean names of tableone
}
