% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regDisplay.R
\name{regDisplay}
\alias{regDisplay}
\title{do regression in batch}
\usage{
regDisplay(
  outcome,
  exposure,
  covariate,
  data,
  regType = list("lm", "glm", "multinom", "coxph", "gee"),
  ...
)
}
\arguments{
\item{outcome}{your outcome. It should only be an element.}

\item{exposure}{your exposure, it could be a vector or a list}

\item{covariate}{your covariate, it coule be a list with vectors}

\item{data}{your data}

\item{...}{other arguments for geepack::glm}
}
\description{
do regression in batch
}
\examples{
# example code
df <- iris
df <- within(df, {
Sepal.Width.2f <- NA_character_
Sepal.Width.2f[Sepal.Width < median(Sepal.Width)] <- "0"
Sepal.Width.2f[Sepal.Width >= median(Sepal.Width)] <- "1"
})

df$Sepal.Width.2f <- factor(df$Sepal.Width.2f, levels = c("0", "1"))
outcome <- "Sepal.Length"
exposure <- c("Sepal.Width", "Sepal.Width.2f")
covariate <- list(c("Petal.Length", "Petal.Width"), c("Species"))
regDisplay(outcome = outcome, exposure = exposure, covariate = covariate, data = df, regType = "lm")


}
\seealso{
[geepack::geeglm()]
}
