% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-lm-tidier.R
\name{regRcs.lm}
\alias{regRcs.lm}
\title{draw RCS}
\usage{
\method{regRcs}{lm}(model, knots = 5, ...)
}
\arguments{
\item{model}{your model}

\item{knots}{the numerb of knots you want to use, default is 5}

\item{...}{Other arguments of rms::Predict function}
}
\value{
a list of objects you might need in RCS plot
\item{fitPred}{This is the dataframe you need when plotting RCS}
\item{pNonLinear}{This is the P for nonlinear}
\item{minimumX}{This is the X of minimum. It is usually used in inflection point detection.}
\item{maximumX}{This is the X of maximum. It is usually used in inflection point detection.}
}
\description{
draw RCS
}
\examples{
# example code

model <- lm("Sepal.Length~ Sepal.Width", data = iris)
c218Tools::regRcs(model = model, knots = 5)

}
\seealso{
[rms::Predict()]
}
