% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbjCnc.R
\name{sbjCnc}
\alias{sbjCnc}
\title{subject conceive}
\usage{
sbjCnc(mesh, st, toClip = FALSE, ...)
}
\arguments{
\item{mesh}{your mesh term}

\item{st}{subject terms}

\item{toClip}{whether to export your result to clipboard}

\item{...}{additional arguments passed to write.table function}
}
\value{
your subjects to use in pubmed
}
\description{
subject conceive
}
\examples{
# examples
# paste mesh
mesh <- "a"
# paste subject terms
st <- "Married Persons
Married Person
Person, Married
Persons, Married
Spouse
Husbands
Husband
Domestic Partners
Domestic Partner
Partner, Domestic
Partners, Domestic
Spousal Notification
Notification, Spousal
Wives
Wife"

sbjCnc(mesh = mesh, st = st)
}
