% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumGLM.R
\name{sumGLM}
\alias{sumGLM}
\title{summary logistic model}
\usage{
sumGLM(
  model,
  n1 = 1,
  n2 = 2,
  latex = T,
  toClip = F,
  pType = "mark",
  desc = F,
  ...
)
}
\arguments{
\item{model}{a linear model object}

\item{n1}{This is the first line you want to get from your summary table. If not specified, it is automatically set when taking your function}

\item{n2}{This is the second line to which you want to get from summary table. If not specified, it is automatically set when taking your function}

\item{latex}{whether to export your result in latex form}

\item{toClip}{whether to export your result to clipboard. This function is still under construction. Don't change it.}

\item{pType}{whether to export your original P, defult to "mark", another option is "value"}

\item{desc}{whether to add description frequency to the final table}

\item{...}{other elements inherited from write.table}
}
\description{
Convert results from linear model. This function take the first variable in independent variables as the target variable.
}
\details{
In academic paper, only one or two lines of regression tables were shown rather than the whole table. Since we are only interested in the specific exposure. Thus, n1 stands for the line started from which we want to extract results. n2 stands for the line to which we want to extract. Normally, you do not need to change them since this package take the first independent variable in your regression model as the variable you are interested in. It will detect which line to take from the final table.
}
\examples{



# example code
input <- mtcars
am.data = glm(formula = am ~ cyl + hp + wt, data = input, family = "binomial")
res <- sumGLM(am.data)
}
