% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geepack-tidier.R
\name{sumReg.geeglm}
\alias{sumReg.geeglm}
\title{sum a(n) geeglm object}
\usage{
\method{sumReg}{geeglm}(
  model,
  n1 = NULL,
  n2 = NULL,
  latex = TRUE,
  toClip = FALSE,
  pType = "mark",
  desc = FALSE,
  digits = 2,
  pDigits = 4,
  ...
)
}
\arguments{
\item{model}{a glm model object}

\item{n1}{This is the first line you want to get from your summary table. If not specified, it is automatically set when taking your function}

\item{n2}{This is the second line to which you want to get from summary table. If not specified, it is automatically set when taking your function}

\item{latex}{whether to export your result in latex form}

\item{toClip}{whether to export your result to clipboard.}

\item{pType}{whether to export your original P, defult to "mark", another option is "value"}

\item{desc}{whether to add description of frequencies between outcome and target exposure variable. default to F}

\item{digits}{digits of beta or OR}

\item{pDigits}{digits of P value}

\item{...}{additional arguments passed to write.table function}
}
\value{
return a tibble of regression table
}
\description{
sumReg summarizes information about the components of a model.
For technical reasons, changing the variable type directly in the formula is not supported. For example:
df <- mtcars
model <- lm("disp ~ as.factor(carb)", data = df)
res <- c218Tools::sumReg(model = model)
The above code chunk would cause error for that the package can not detect the target variable levels.
Rather, Please do it before the model.
df <- mtcars
df$carb <- as.factor(df$carb)
model <- lm("disp~ carb", data = df)
res <- c218Tools::sumReg(model = model)
}
\details{
In academic paper, only one or two lines of regression tables were shown rather than the whole table. Since we are only interested in the specific exposure. Thus, n1 stands for the line started from which we want to extract results. n2 stands for the line to which we want to extract. Normally, you do not need to change them since this package take the first independent variable in your regression model as the variable you are interested in. It will detect which line to take from the final table.
}
\examples{
library(tidyverse)
data(package = "geepack", dietox)
dietox$Cu <- as.factor(dietox$Cu)
mf <- formula(Weight ~ Cu * (Time + I(Time^2) + I(Time^3))) \%>\% as.formula
gee1 <- geepack::geeglm(mf, data=dietox, id=Pig, family=poisson("identity"), corstr="ar1")
res <- sumReg(gee1)

}
