% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable-format-tidier.R
\name{themeGm}
\alias{themeGm}
\title{This is}
\usage{
themeGm(x)
}
\arguments{
\item{x}{a flextable object}
}
\value{
return a flextable that meets the need of academic paper
}
\description{
This is
}
\details{
In academic paper, we use three line table. This is a simpilified version to export the table to save time for group meeting.
}
\examples{
# examples


iris \%>\%
  flextable() \%>\%
  themeGm()
}
