% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c218Tools-package.R
\docType{package}
\name{c218Tools-package}
\alias{c218Tools}
\alias{c218Tools-package}
\title{c218Tools: Summary the regression table}
\description{
This package tidy the table into academic paper format.
}
\keyword{internal}
