% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTableOne.R
\name{changeLevelTwoFactor}
\alias{changeLevelTwoFactor}
\title{sum a(n) tableone object}
\usage{
changeLevelTwoFactor(df, outcomeLevel = NULL)
}
\arguments{
\item{df}{a tableone object}

\item{outcomeLevel}{level of outcome if it were categorical}
}
\value{
return a tibble of regression table
}
\description{
sumReg summarizes information about the components of a model.
For technical reasons, changing the variable type directly in the formula is not supported. For example:
df <- mtcars
model <- lm("disp ~ as.factor(carb)", data = df)
res <- c218Tools::sumReg(model = model)
The above code chunk would cause error for that the package can not detect the target variable levels.
Rather, Please do it before the model.
df <- mtcars
df$carb <- as.factor(df$carb)
model <- lm("disp~ carb", data = df)
res <- c218Tools::sumReg(model = model)
}
\details{
In academic paper, only one or two lines of regression tables were shown rather than the whole table. Since we are only interested in the specific exposure. Thus, n1 stands for the line started from which we want to extract results. n2 stands for the line to which we want to extract. Normally, you do not need to change them since this package take the first independent variable in your regression model as the variable you are interested in. It will detect which line to take from the final table.
}
\examples{
# example code

df <- iris \%>\%
  dplyr::mutate(
    Petal.Length.2f = case_when(
      Petal.Length < 2 ~ "small",
      Petal.Length >= 2 ~ "large",
      TRUE ~ NA_character_
    )
  )
table1 <- tableone::CreateTableOne(vars = c("Sepal.Length", "Sepal.Width", "Petal.Length.2f"), strata = "Species", data = iris) \%>\%
  print(showAllLevels = T)

res <- c218Tools::sumTableOne(table1)
}
