% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTargetLevels.R
\name{detectTargetLevels}
\alias{detectTargetLevels}
\title{detect target levels}
\usage{
detectTargetLevels(target, data)
}
\arguments{
\item{target}{your target exposure}

\item{data}{your data}
}
\description{
detect target levels
}
