% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable-format-tidier.R
\name{formatTable}
\alias{formatTable}
\title{Function To set table style}
\usage{
formatTable(x)
}
\arguments{
\item{x}{a flextable object}
}
\value{
return a flextable that meets the need of academic paper
}
\description{
Function To set table style
}
\details{
In academic paper, we use three line table. This is a simpilified version to export the table to save time for group meeting.
}
\examples{
# examples


iris \%>\%
  flextable() \%>\%
  themeGm()
}
