% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumLM.R
\name{sumLM}
\alias{sumLM}
\title{summary linear model}
\usage{
sumLM(form, n1 = 1, n2 = 2, data, latex = T, toClip = F, pType = "mark", ...)
}
\arguments{
\item{form}{a linear model object}

\item{n1}{This is the first line you want to get from your summary table}

\item{n2}{This is the second line to which you want to get from summary table}

\item{data}{your data}

\item{latex}{whether to export your result in latex form}

\item{toClip}{whether to export your result to clipboard}

\item{pType}{whether to export your original P, defult to "mark", another option is "value"}

\item{...}{other element inherited from write.table function}
}
\description{
Convert results from linear model
}
