% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTableOne.R
\name{sumTableOne}
\alias{sumTableOne}
\title{summary table one from tableone::CreateTableOne function}
\usage{
sumTableOne(tableOne, latex = F, toClip = F, pType = "mark", ...)
}
\arguments{
\item{...}{other elements inherited from write.table}

\item{model}{a tableone object}
}
\description{
Convert results from tableone::CreateTableOne
}
\details{
In academic paper, only one or two lines of regression tables were shown rather than the whole table. Since we are only interested in the specific exposure. Thus, n1 stands for the line started from which we want to extract results. n2 stands for the line to which we want to extract. Normally, you do not need to change them since this package take the first independent variable in your regression model as the variable you are interested in. It will detect which line to take from the final table.
}
\examples{


# example code
model <- lm("Sepal.Length~ Sepal.Width", data = iris)
regTab <- c218Tools::sumLM(model = model, toClip = TRUE)
}
