% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTableOne.R
\name{sumTableOne}
\alias{sumTableOne}
\title{summary table one from tableone::CreateTableOne function}
\usage{
sumTableOne(tableOne, toClip = F, latex = F, ...)
}
\arguments{
\item{tableOne}{a tableone object}

\item{toClip}{whether to export your result to clipboard. When you are using this package on remote server, This paramater would not work for that linux does not provide access to clipboard when you use ssh to log in.}

\item{latex}{whether to export your result in latex form.In this package, it was set default to \verb{\\ast}, \verb{\\dag}, \verb{\\ddag}. to replace * $ # respectively. For more information about latex, please refer to https://en.wikibooks.org/wiki/LaTeX/Mathematics}

\item{...}{other elements inherited from write.table}
}
\description{
summary table one from tableone::CreateTableOne function
}
\details{
When using this function. Remember to use `showAllLevels = T` in the print function from tableone package. Also, make sure you have at least one continuous variable and one categoriacal variable to make the function run.
}
\examples{
# example code

df <- iris \%>\%
  dplyr::mutate(
    Petal.Length.2f = case_when(
      Petal.Length < 2 ~ "small",
      Petal.Length >= 2 ~ "large",
      TRUE ~ NA_character_
    )
  )
table1 <- tableone::CreateTableOne(vars = c("Sepal.Length", "Sepal.Width", "Petal.Length.2f"), strata = "Species", data = iris) \%>\%
  print(showAllLevels = T)

res <- c218Tools::sumTableOne(table1)
}
\seealso{
[utils::write.table()]
}
